move progpow configs to algorithm
